Fixes circular import
