+++ bookwyrm/connectors/inventaire.py
@@ -2,7 +2,8 @@
 import re
 
 from bookwyrm import models
-from .abstract_connector import AbstractConnector, SearchResult, Mapping
+from bookwyrm.book_search import SearchResult
+from .abstract_connector import AbstractConnector, Mapping
 from .abstract_connector import get_data
 from .connector_manager import ConnectorException
